#24 - chore: 소비성향 관련 클래스 core 모듈로 이동
core 모듈에서 성향 정보를 바탕으로 id를 지정해주기 위해 core로 이동
