+++ generator-core/src/main/resources/preference.json
