+++ generator-simulator/src/test/java/com/simpaylog/generatorsimulator/util/ConsumptionDeltaAllocatorTest.java
@@ -3,10 +3,10 @@ package com.simpaylog.generatorsimulator.util;
 import com.simpaylog.generatorcore.cache.IncomeLevelLocalCache;
 import com.simpaylog.generatorcore.cache.dto.IncomeLevelInfo;
 import com.simpaylog.generatorsimulator.TestConfig;
-import com.simpaylog.generatorsimulator.configuration.PreferenceLocalCache;
-import com.simpaylog.generatorsimulator.dto.ConsumptionDelta;
-import com.simpaylog.generatorsimulator.dto.MonthlyConsumptionCost;
-import com.simpaylog.generatorsimulator.dto.PreferenceInfos;
+import com.simpaylog.generatorcore.cache.PreferenceLocalCache;
+import com.simpaylog.generatorcore.cache.dto.preference.ConsumptionDelta;
+import com.simpaylog.generatorcore.cache.dto.preference.MonthlyConsumptionCost;
+import com.simpaylog.generatorcore.cache.dto.preference.PreferenceInfos;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
